RateMan/requirements.txt
codecov
dash
dask
matplotlib
numpy
pandas
paramiko
pint==0.9
pycodestyle
scipy
six
sphinx
sphinx_rtd_theme
tornado
python-telegram-bot
